Freeze

+++ src/Location.java
@@ -1,17 +1,17 @@
 
 public class Location {
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 	
-	public Location(int x, int y)
+	public Location(double x, double y)
 	{
 		this.x = x;
 		this.y = y;
 	}
 	
-	public int getX(){return x;}
-	public int getY() {return y;}
-	public void incX(int x) {this.x += x;} //increments x by given value in param
-	public void incY(int y) {this.y += y;}
+	public double getX(){return x;}
+	public double getY() {return y;}
+	public void incX(double x) {this.x += x;} //increments x by given value in param
+	public void incY(double y) {this.y += y;}
 	
 }
